Add files via upload
This model written in Python using Mesa3 aims to reproduce the model xnx_imitate_if_better.nlogo. The model should work quite well; you can modify the payoff matrix, the revision probability, and the number of agents and steps to obtain different results. In the end, the model should produce a stacked histogram where the model's evolution is visualized.