Update forest_fire.py
Not perfect. It gives me some warnings related to agent placement, but the model still works. The code only shows the final state of the forest (trees that are still alive in green, or burned in red) and the graph showing the percentage of burned trees over time.
Set density at 60%